Made the code to not have serialization errors
