+++ e2/src/test/scala/io/prediction/e2/evaluation/CrossValidationTest.scala
@@ -1,6 +1,6 @@
 package io.prediction.e2.evaluation
 
-import io.prediction.e2.fixture.{CrossValidationFixture, SharedSparkContext}
+import io.prediction.e2.fixture.SharedSparkContext
 import org.scalatest.{Matchers, FlatSpec}
 
 import io.prediction.e2.engine.LabeledPoint
@@ -9,30 +9,54 @@ import org.apache.spark.rdd.RDD
 
 import scala.language.reflectiveCalls
 
+object CrossValidationTest {
+  case class TrainingData(labeledPoints: Seq[LabeledPoint])
+  case class Query(features: Array[String])
+  case class ActualResult(label: String)
+
+  def toTrainingData(labeledPoints: RDD[LabeledPoint]) = TrainingData(labeledPoints.collect().toSeq)
+  def toQuery(labeledPoint: LabeledPoint) = Query(labeledPoint.features)
+  def toActualResult(labeledPoint: LabeledPoint) = ActualResult(labeledPoint.label)
+
+}
+
 class CrossValidationTest extends FlatSpec with Matchers
-with SharedSparkContext with CrossValidationFixture {
-  val labeledPoints = dataset.labeledPoints
-  val dataCount = labeledPoints.length
+with SharedSparkContext{
+
+
+  val Label1 = "l1"
+  val Label2 = "l2"
+  val Label3 = "l3"
+  val Attribute1 = "a1"
+  val NotAttribute1 = "na1"
+  val Attribute2 = "a2"
+  val NotAttribute2 = "na2"
+
+  val labeledPoints = Seq(
+    LabeledPoint(Label1, Array(Attribute1, Attribute2)),
+    LabeledPoint(Label2, Array(NotAttribute1, Attribute2)),
+    LabeledPoint(Label3, Array(Attribute1, NotAttribute2))
+  )
+
+  val dataCount = labeledPoints.size
   val evalKs = (1 to dataCount)
   type D = LabeledPoint
-  type TD = TrainingData
+  type TD = CrossValidationTest.TrainingData
   type EI = Null
-  type Q = Query
-  type A = ActualResult
+  type Q = CrossValidationTest.Query
+  type A = CrossValidationTest.ActualResult
   type DataSplit = (TD, EI, RDD[(Q, A)])
 
   "Fold count" should "equal evalK" in {
-    evalKs.map {k =>
       val labeledPointsRDD = sc.parallelize(labeledPoints)
       val splits = CommonHelperFunctions.splitData[D, TD, EI, Q, A](
-        k,
+        dataCount,
         labeledPointsRDD,
         null,
-        toTrainingData,
-        toQuery,
-        toActualResult)
-      splits.length should be (k)
-    }
+        CrossValidationTest.toTrainingData,
+        CrossValidationTest.toQuery,
+        CrossValidationTest.toActualResult)
+      splits.size shouldEqual dataCount
   }
 
 
@@ -43,9 +67,9 @@ with SharedSparkContext with CrossValidationFixture {
         k,
         labeledPointsRDD,
         null,
-        toTrainingData,
-        toQuery,
-        toActualResult)
+        CrossValidationTest.toTrainingData,
+        CrossValidationTest.toQuery,
+        CrossValidationTest.toActualResult)
       splits.map {split: DataSplit =>
         val foldSize = dataCount / k
         if (dataCount % k == 0) {
@@ -66,9 +90,9 @@ with SharedSparkContext with CrossValidationFixture {
         k,
         labeledPointsRDD,
         null,
-        toTrainingData,
-        toQuery,
-        toActualResult)
+        CrossValidationTest.toTrainingData,
+        CrossValidationTest.toQuery,
+        CrossValidationTest.toActualResult)
       splits.map {split: DataSplit =>
         val trainginDataRDD = sc.parallelize(split._1.labeledPoints)
         val collected = trainginDataRDD.union(split._2).collect()
